extract PlayerMarshallerService.groovy
